JIT-125423 make sure to convert flot to int

+++ src/ocp/awesome_mapper2/functions.go
@@ -83,12 +83,12 @@ func init() {
 	AddFunction("int", func(args ...interface{}) (interface{}, error) {
 		switch t := args[0].(type) {
 		case int, int8, int16, int32, int64, uint, uint8, uint16, uint32, uint64, uintptr:
-			return float64(reflect.ValueOf(t).Int()), nil
+			return int(reflect.ValueOf(t).Int()), nil
 		case float32, float64:
-			return float64(reflect.ValueOf(t).Float()), nil
+			return int(reflect.ValueOf(t).Float()), nil
 		case string:
 			float, err := strconv.ParseFloat(t, 64)
-			return float, err
+			return int(float), err
 		default:
 			return nil, errors.New("cant parse non-string")
 		}
